refactoring to see whether dask cant recognise function handle
